chore(engine): create rmq worker
